training-log: fix bug with newline character
Putting a \n in a string made Goldencheetah either save it as an actual
newline or read it in as an actual newline. Whichever it is, it leads to
python syntax breaking an error gets thrown. Instead use chr() to turn
the ASCII number into the character and then use that whereever needed.

To be exact, this

```
    "\n"
```

turns into

```
    "
"
```