rolling: adjust for exec prior to today's run
Without it, you get a little dip at the end since it compares today's 0
km to whatever it was x days prior. That doesn't mean much though if you
still have to run later in the day. Instead make it pretend to be
yesterday and not generate anything for today yet.

This adjustment only happens if today's counter is at 0 still. Otherwise
it is assumed you already ran.