Adjust activity datetime based on keyword
Specifically a tz:utc±N keyword. Add the hours to the activity's start
in the diary. I'll see later if I care enough to just add an extra
property for an adjust time instead.

In diary it was most notable because of the day-per-day view. I guess in
some overview graphs it might too, but perhaps less so since usually per
week.